requirements.txt
# Core ML and NLP
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
pandas==2.1.4
numpy==1.24.3

# PDF Processing
PyPDF2==3.0.1
python-docx==1.0.1

# Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.11.0

# Utilities
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1 